+++ src/Ui/BotListPage/BotWidget/BotSetupPage.py
@@ -14,7 +14,7 @@ from PySide6.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout
 # 项目内模块导入
 from src.Ui.AddPage.Connect import ConnectWidget
 from src.Ui.AddPage.Advanced import AdvancedWidget
-from src.Ui.AddPage.BotWidget import BotWidget
+from src.Ui.AddPage.bot_widget import BotWidget
 from src.Core.Config.ConfigModel import Config
 
 if TYPE_CHECKING:
